Changes from moving IndexSortKey
