+++ mongo/src/test/java/com/redhat/lightblue/mongo/crud/MongoCRUDControllerTest.java
@@ -37,9 +37,9 @@ import com.mongodb.DBCollection;
 import com.mongodb.DBCursor;
 import com.mongodb.DBObject;
 import com.redhat.lightblue.crud.CRUDDeleteResponse;
+import com.redhat.lightblue.crud.CRUDFindResponse;
 import com.redhat.lightblue.crud.CRUDInsertionResponse;
 import com.redhat.lightblue.crud.CRUDOperation;
-import com.redhat.lightblue.crud.CRUDFindResponse;
 import com.redhat.lightblue.crud.CRUDSaveResponse;
 import com.redhat.lightblue.crud.CRUDUpdateResponse;
 import com.redhat.lightblue.crud.DocCtx;
@@ -48,6 +48,7 @@ import com.redhat.lightblue.metadata.EntityMetadata;
 import com.redhat.lightblue.metadata.FieldConstraint;
 import com.redhat.lightblue.metadata.FieldCursor;
 import com.redhat.lightblue.metadata.Index;
+import com.redhat.lightblue.metadata.IndexSortKey;
 import com.redhat.lightblue.metadata.MetadataStatus;
 import com.redhat.lightblue.metadata.ObjectField;
 import com.redhat.lightblue.metadata.SimpleField;
@@ -58,11 +59,6 @@ import com.redhat.lightblue.metadata.types.StringType;
 import com.redhat.lightblue.mongo.common.DBResolver;
 import com.redhat.lightblue.mongo.common.MongoDataStore;
 import com.redhat.lightblue.mongo.config.MongoConfiguration;
-import com.redhat.lightblue.mongo.crud.IterateAndUpdate;
-import com.redhat.lightblue.mongo.crud.MongoCRUDController;
-import com.redhat.lightblue.mongo.crud.MongoCrudConstants;
-import com.redhat.lightblue.mongo.crud.Translator;
-import com.redhat.lightblue.query.IndexSortKey;
 import com.redhat.lightblue.query.Projection;
 import com.redhat.lightblue.util.Error;
 import com.redhat.lightblue.util.JsonDoc;
